rename cloudfronthelpers package to cloudfront
